+++ shuup_workbench/settings/base_settings.py
@@ -54,6 +54,7 @@ INSTALLED_APPS = add_enabled_addons(SHUUP_ENABLED_ADDONS_FILE, [
     'shuup.utils',
     'shuup.xtheme',
     'shuup.reports',
+    'shuup.default_reports',
     # external apps
     'bootstrap3',
     'django_jinja',